refactor(home): 홈페이지 라우트 이동 및 구조 리팩토링
 홈페이지 라우트를 app 디렉토리로 이동
 홈페이지 UI 컴포넌트를 screens/home/ui로 이동
 홈페이지 레이아웃을 screens/home/ui/RootLayout으로 이동
 홈페이지 스타일을 app/styles/globals.css로 이동
 홈페이지 컴포넌트들 이동 (MainNews, MainShortforms, Top3, Top3Posts 등)
 동적 경로 상수 사용 (DYNAMIC_PATH)
 홈페이지 레이아웃 구조 개선 (MobileHeader 추가)
