+++ apps/web/app/layout.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/screens/home/ui/RootLayout';
