+++ apps/web/app/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/screens/home/ui/HomePage';
